Upgrade to manifest v3 (#44)
- separate background and injected script
- remove `webextension-polyfill` as all APIs are promise based in MV3
- upgrade `turndown` to not depdend on `jsdom`
- upgrade some dev dependencies
