+++ package.json
@@ -13,27 +13,26 @@
 		"release": "VER=$(daily-version) run-s build version release:*"
 	},
 	"devDependencies": {
-		"@types/firefox-webext-browser": "^70.0.1",
+		"@types/firefox-webext-browser": "^120.0.3",
 		"chrome-webstore-upload-cli": "^1.2.0",
-		"copy-webpack-plugin": "^5.1.1",
+		"copy-webpack-plugin": "^12.0.2",
 		"daily-version": "^1.0.0",
 		"dot-json": "^1.2.0",
 		"eslint": "^6.8.0",
 		"eslint-config-xo": "^0.26.0",
 		"npm-run-all": "^4.1.5",
 		"size-plugin": "^2.0.1",
-		"terser-webpack-plugin": "^2.3.5",
+		"terser-webpack-plugin": "^5.3.10",
 		"utc-version": "^2.0.1",
 		"web-ext": "^4.1.0",
 		"web-ext-submit": "^4.1.0",
-		"webpack": "^4.42.1",
-		"webpack-cli": "^3.3.11",
+		"webpack": "^5.92.0",
+		"webpack-cli": "^5.1.4",
 		"xo": "^0.28.1"
 	},
 	"dependencies": {
-		"turndown": "^6.0.0",
-		"turndown-plugin-gfm": "^1.0.2",
-		"webextension-polyfill": "^0.4.0"
+		"turndown": "^7.2.0",
+		"turndown-plugin-gfm": "^1.0.2"
 	},
 	"xo": {
 		"envs": [
@@ -43,7 +42,9 @@
 			"distribution"
 		],
 		"globals": [
-			"browser"
+			"browser",
+			"chrome",
+			"globalThis"
 		],
 		"rules": {
 			"unicorn/string-content": "off"
